refactor: extract arm contract into rover_arm_common so HMIs/driver decouple
Move the shared arm contract headers (motor_addressing, motor_config,
arm_commands, arm_telemetry, ArmSerialProtocol) into a new header-only,
Qt-free rover_arm_common package. arm_hardware_interface no longer links
the Qt GUI lib, and arm_control/rover_hmi no longer depend on the driver.

rover_hmi_core and arm_hardware_interface are now dependency leaves, so
either can be excluded from colcon builds without breaking downstream.
