+++ src/arm_control/package.xml
@@ -40,7 +40,7 @@
   <depend>rover_msgs</depend>
   <depend>rover_utils</depend> 
   <depend>moteus_msgs</depend>
-  <depend>arm_hardware_interface</depend>
+  <depend>rover_arm_common</depend>
 
 
   <test_depend>ament_cmake_gtest</test_depend>
