+++ src/arm_hardware_interface/include/moteus_driver_node.h
@@ -50,10 +50,10 @@
 #include "std_msgs/msg/string.hpp"
 #include "sensor_msgs/msg/joint_state.hpp"
 
-// #include <rover_hmi_core/arm/motor_addressing.h>   // NUM_MOTORS, ARM_JOINTS, unit converters
-#include <rover_hmi_core/arm/motor_config.h>       // MotorConfig, get_arm_configuration()
-#include <rover_hmi_core/arm/arm_commands.h>       // CMD_*, MotorCommand
-#include <rover_hmi_core/arm/arm_telemetry.h>      // MotorTelem
+// #include <rover_arm_common/motor_addressing.h>   // NUM_MOTORS, ARM_JOINTS, unit converters
+#include <rover_arm_common/motor_config.h>       // MotorConfig, get_arm_configuration()
+#include <rover_arm_common/arm_commands.h>       // CMD_*, MotorCommand
+#include <rover_arm_common/arm_telemetry.h>      // MotorTelem
 #include "moteus_protocol.h"                       // MoteusProtocol::make*Frame(), parseReply()
