+++ src/rover_arm_common/package.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>rover_arm_common</name>
+  <version>0.1.0</version>
+  <description>
+    Shared arm contract: motor addressing, motor configuration, command and
+    telemetry structs, and the legacy serial protocol constants. Header-only
+    and Qt-free so the arm driver, arm control, and both HMIs can share one
+    definition without depending on each other.
+  </description>
+  <maintainer email="rhimaaron@gmail.com">Aaron Rhim</maintainer>
+  <license>MIT</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
