+++ src/rover_hmi_core/include/rover_hmi_core/arm/motor_config_module.h
@@ -27,7 +27,7 @@
 #pragma once
 
 #include <rover_hmi_core/gui_module.h>
-#include "motor_addressing.h"
+#include <rover_arm_common/motor_addressing.h>
 
 #include <QLineEdit>
 #include <QLabel>
