+++ src/rover_hmi_core/include/rover_hmi_core/arm/motor_status_module.h
@@ -13,7 +13,7 @@
 #pragma once
 
 #include <rover_hmi_core/gui_module.h>
-#include "motor_addressing.h"  // NUM_MOTORS
+#include <rover_arm_common/motor_addressing.h>  // NUM_MOTORS
 
 #include <QLabel>
 #include <QScrollArea>
