+++ src/rover_hmi_core/package.xml
@@ -16,6 +16,7 @@
   <depend>sensor_msgs</depend>
   <depend>geometry_msgs</depend>
   <depend>ament_index_cpp</depend>
+  <depend>rover_arm_common</depend>
 
   <build_depend>qt5-qmake</build_depend>
   <build_depend>qtbase5-dev</build_depend>
